matrices: add sympy.matrices.exceptions module
