+++ sympy/matrices/__init__.py
@@ -3,7 +3,8 @@
 Includes functions for fast creating matrices like zero, one/eye, random
 matrix, etc.
 """
-from .common import ShapeError, NonSquareMatrixError, MatrixKind
+from .exceptions import ShapeError, NonSquareMatrixError
+from .common import MatrixKind
 from .dense import (
     GramSchmidt, casoratian, diag, eye, hessian, jordan_cell,
     list2numpy, matrix2numpy, matrix_multiply_elementwise, ones,
