+++ sympy/matrices/decompositions.py
@@ -5,7 +5,7 @@ from sympy.core.function import expand_mul
 from sympy.functions.elementary.miscellaneous import Min, sqrt
 from sympy.functions.elementary.complexes import sign
 
-from .common import NonSquareMatrixError, NonPositiveDefiniteMatrixError
+from .exceptions import NonSquareMatrixError, NonPositiveDefiniteMatrixError
 from .utilities import _get_intermediate_simp, _iszero
 from .determinant import _find_reasonable_pivot_naive
