+++ sympy/matrices/determinant.py
@@ -10,7 +10,7 @@ from sympy.functions.combinatorial.numbers import nC
 from sympy.polys.matrices.domainmatrix import DomainMatrix
 from sympy.polys.matrices.ddm import DDM
 
-from .common import NonSquareMatrixError
+from .exceptions import NonSquareMatrixError
 from .utilities import (
     _get_intermediate_simp, _get_intermediate_simp_bool,
     _iszero, _is_zero_after_expand_mul, _dotprodsimp, _simplify)
