+++ sympy/solvers/solvers.py
@@ -44,7 +44,7 @@ from sympy.simplify import (simplify, collect, powsimp, posify,  # type: ignore
 from sympy.simplify.sqrtdenest import sqrt_depth
 from sympy.simplify.fu import TR1, TR2i, TR10, TR11
 from sympy.strategies.rl import rebuild
-from sympy.matrices.common import NonInvertibleMatrixError
+from sympy.matrices.exceptions import NonInvertibleMatrixError
 from sympy.matrices import Matrix, zeros
 from sympy.polys import roots, cancel, factor, Poly
 from sympy.polys.solvers import sympy_eqs_to_ring, solve_lin_sys
